made changes in the cycles algo
The algorithm for producing all the cycles in an undirected graph has been successfully shaped and finalised. The algorithm works perfectly fine for all kind of graphs with minimum 1 cycle or even no cycle in which case it will not print anything. Algorithm uses the concept of visited and parent araay extensively. The main game-changing function is color_func. It sorts the order of any cycle and make it ready for comparision with further cycles to be inserted in the color array.